+++ lib/puppet/provider/pingdom_check/pop3.rb
@@ -1,3 +1,5 @@
+require File.expand_path(File.join(File.dirname(__FILE__), 'check_base.rb'))
+
 Puppet::Type.type(:pingdom_check).provide(:pop3, :parent => :check_base) do
     has_features :port, :stringtoexpect, :encryption
