+++ lib/puppet/provider/pingdom_check/tcp.rb
@@ -1,3 +1,5 @@
+require File.expand_path(File.join(File.dirname(__FILE__), 'check_base.rb'))
+
 Puppet::Type.type(:pingdom_check).provide(:tcp, :parent => :check_base) do
     has_features :port, :stringtosend, :stringtoexpect
